+++ layers/Schema/packages/media/src/FieldResolvers/ObjectType/MediaObjectTypeFieldResolver.php
@@ -15,7 +15,7 @@ use PoP\Engine\TypeResolvers\ScalarType\StringScalarTypeResolver;
 use PoPSchema\Media\TypeAPIs\MediaTypeAPIInterface;
 use PoPSchema\Media\TypeResolvers\ObjectType\MediaObjectTypeResolver;
 use PoPSchema\SchemaCommons\ModuleProcessors\CommonFilterInputContainerModuleProcessor;
-use PoPSchema\SchemaCommons\TypeResolvers\ScalarType\DateScalarTypeResolver;
+use PoPSchema\SchemaCommons\TypeResolvers\ScalarType\DateTimeScalarTypeResolver;
 use PoPSchema\SchemaCommons\TypeResolvers\ScalarType\URLScalarTypeResolver;
 
 class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolver
@@ -25,7 +25,7 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
     private ?URLScalarTypeResolver $urlScalarTypeResolver = null;
     private ?IntScalarTypeResolver $intScalarTypeResolver = null;
     private ?StringScalarTypeResolver $stringScalarTypeResolver = null;
-    private ?DateScalarTypeResolver $dateScalarTypeResolver = null;
+    private ?DateTimeScalarTypeResolver $dateTimeScalarTypeResolver = null;
 
     final public function setMediaTypeAPI(MediaTypeAPIInterface $mediaTypeAPI): void
     {
@@ -67,13 +67,13 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
     {
         return $this->stringScalarTypeResolver ??= $this->instanceManager->getInstance(StringScalarTypeResolver::class);
     }
-    final public function setDateScalarTypeResolver(DateScalarTypeResolver $dateScalarTypeResolver): void
+    final public function setDateTimeScalarTypeResolver(DateTimeScalarTypeResolver $dateTimeScalarTypeResolver): void
     {
-        $this->dateScalarTypeResolver = $dateScalarTypeResolver;
+        $this->dateTimeScalarTypeResolver = $dateTimeScalarTypeResolver;
     }
-    final protected function getDateScalarTypeResolver(): DateScalarTypeResolver
+    final protected function getDateTimeScalarTypeResolver(): DateTimeScalarTypeResolver
     {
-        return $this->dateScalarTypeResolver ??= $this->instanceManager->getInstance(DateScalarTypeResolver::class);
+        return $this->dateTimeScalarTypeResolver ??= $this->instanceManager->getInstance(DateTimeScalarTypeResolver::class);
     }
 
     public function getObjectTypeResolverClassesToAttachTo(): array
@@ -96,7 +96,9 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
             'altText',
             'description',
             'date',
+            'dateAsString',
             'modified',
+            'modifiedAsString',
             'mimeType',
         ];
     }
@@ -113,8 +115,10 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
             'caption' => $this->getStringScalarTypeResolver(),
             'altText' => $this->getStringScalarTypeResolver(),
             'description' => $this->getStringScalarTypeResolver(),
-            'date' => $this->getDateScalarTypeResolver(),
-            'modified' => $this->getDateScalarTypeResolver(),
+            'date' => $this->getDateTimeScalarTypeResolver(),
+            'dateAsString' => $this->getStringScalarTypeResolver(),
+            'modified' => $this->getDateTimeScalarTypeResolver(),
+            'modifiedAsString' => $this->getStringScalarTypeResolver(),
             'mimeType' => $this->getStringScalarTypeResolver(),
             default => parent::getFieldTypeResolver($objectTypeResolver, $fieldName),
         };
@@ -125,7 +129,9 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
         return match ($fieldName) {
             'src',
             'date',
-            'modified'
+            'dateAsString',
+            'modified',
+            'modifiedAsString'
                 => SchemaTypeModifiers::NON_NULLABLE,
             default
                 => parent::getFieldTypeModifiers($objectTypeResolver, $fieldName),
@@ -145,7 +151,9 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
             'altText' => $this->getTranslationAPI()->__('Media element alt text', 'pop-media'),
             'description' => $this->getTranslationAPI()->__('Media element description', 'pop-media'),
             'date' => $this->getTranslationAPI()->__('Media element\'s published date', 'pop-media'),
+            'dateAsString' => $this->getTranslationAPI()->__('Media element\'s published date, in String format', 'pop-media'),
             'modified' => $this->getTranslationAPI()->__('Media element\'s modified date', 'pop-media'),
+            'modifiedAsString' => $this->getTranslationAPI()->__('Media element\'s modified date, in String format', 'pop-media'),
             'mimeType' => $this->getTranslationAPI()->__('Media element\'s mime type', 'pop-media'),
             default => parent::getFieldDescription($objectTypeResolver, $fieldName),
         };
@@ -178,8 +186,10 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
     public function getFieldFilterInputContainerModule(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName): ?array
     {
         return match ($fieldName) {
-            'date' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
-            'modified' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
+            'date' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE],
+            'dateAsString' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
+            'modified' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE],
+            'modifiedAsString' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
             default => parent::getFieldFilterInputContainerModule($objectTypeResolver, $fieldName),
         };
     }
@@ -227,15 +237,19 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
             case 'description':
                 return $this->getMediaTypeAPI()->getDescription($media);
             case 'date':
-                return new DateTime($this->getDateFormatter()->format(
+                return new DateTime($this->getMediaTypeAPI()->getDate($media, $fieldArgs['gmt']));
+            case 'dateAsString':
+                return $this->getDateFormatter()->format(
                     $fieldArgs['format'],
                     $this->getMediaTypeAPI()->getDate($media, $fieldArgs['gmt'])
-                ));
+                );
             case 'modified':
-                return new DateTime($this->getDateFormatter()->format(
+                return new DateTime($this->getMediaTypeAPI()->getModified($media, $fieldArgs['gmt']));
+            case 'modifiedAsString':
+                return $this->getDateFormatter()->format(
                     $fieldArgs['format'],
                     $this->getMediaTypeAPI()->getModified($media, $fieldArgs['gmt'])
-                ));
+                );
             case 'mimeType':
                 return $this->getMediaTypeAPI()->getMimeType($media);
         }